DISCOVERYACCESS-7911 - merge with dev - new/modified files

+++ app/controllers/kpanel_controller.rb
@@ -7,8 +7,7 @@ class KpanelController < ApplicationController
   #attr_accessible :authq, :start, :order, :browse_type
   @@browse_index_author = ENV['BROWSE_INDEX_AUTHOR'].nil? ? 'author' : ENV['BROWSE_INDEX_AUTHOR']
   @@browse_index_subject = ENV['BROWSE_INDEX_SUBJECT'].nil? ? 'subject' : ENV['BROWSE_INDEX_SUBJECT']
-  @@browse_author = Blacklight.connection_config[:url].gsub(/\/solr\/.*/,'/solr/') + @@browse_index_author
-  @@browse_subject = Blacklight.connection_config[:url].gsub(/\/solr\/.*/,'/solr/') + @@browse_index_subject
+  @@browse_index_authortitle = ENV['BROWSE_INDEX_AUTHORTITLE'].nil? ? 'authortitle' : ENV['BROWSE_INDEX_AUTHORTITLE']
 
   def heading
    @heading='Browse'
@@ -19,8 +18,8 @@ class KpanelController < ApplicationController
   	result = nil
     if(auth_type != "" && authq != "")
       p =  {"q" => authq.gsub("\\"," ").gsub("%26", "&")}
-      base_url = (auth_type == "author") ? @@browse_author: @@browse_subject
-      query_url = base_url + "/browse?wt=json&" + p.to_param
+      base_solr_url = Blacklight.connection_config[:url].gsub(/\/solr\/.*/,'/solr/')
+      query_url = base_solr_url + browse_type(auth_type) + '/browse?wt=json&' + p.to_param
       url = URI.parse(query_url)
       resp = Net::HTTP.get_response(url)
       data = resp.body
@@ -35,23 +34,41 @@ class KpanelController < ApplicationController
   # Get info to display in knowledge panel
   # Arguments: base url, authority type (author or subject), query string 
   def panel
-  	@authq = params[:authq] || ""
-  	@auth_type = params[:type] || ""
-  	# This should return the JSON for solr documents
+    return unless ['author', 'subject', 'authortitle'].include?(params[:type])
+
+    @authq = params[:authq] || ''
+    @auth_type = params[:type] || ''
+    @bib = params[:bib] || ''
+    # This should return the JSON for solr documents
     results = get_browse_info(@authq, @auth_type)
     @data = results[0]
-    heading = @data["heading"]
-    #Call helper method to see whether external data should or should not be included within panel
-    #This used to be @authq but changing it because the parameter from the link may have different
-    #ending punctuation than the heading in the browse.  Relying entirely on heading now
-    permissions_key = heading.delete_prefix('"').delete_suffix('"')
-    @exclusions = check_permissions(permissions_key)
-    render "panel"
-    
-  end	
+
+    if ['author', 'subject'].include?(@auth_type)
+      heading = @data['heading']
+      #Call helper method to see whether external data should or should not be included within panel
+      #This used to be @authq but changing it because the parameter from the link may have different
+      #ending punctuation than the heading in the browse.  Relying entirely on heading now
+      permissions_key = heading.delete_prefix('"').delete_suffix('"')
+      @exclusions = check_permissions(permissions_key)
+      render 'default_panel'
+    elsif @auth_type == 'authortitle'
+      render 'author_title_panel'
+    end
+  end
   def index
   	
   end
 
- 
+  private
+
+  def browse_type(auth_type)
+    case auth_type
+    when 'author'
+      @@browse_index_author
+    when 'subject'
+      @@browse_index_subject
+    when 'authortitle'
+      @@browse_index_authortitle
+    end
+  end
 end
